### 2024-01-29 #6250 @AnjaliSandip
:wave: @akabla, @Datseris & @SotaYoshida, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
