### 2024-01-31 #6250 @akabla
Many thanks for the invitation. I'm a bit tight at the moment, but I would warmly recommend Fikret Basar @fb456 as a reviewer. Fikret currently is a graduate student in my group, knowledgeable in Mechanics and Julia, and possibly author of the JOSS paper in the next few months.
