### 2024-02-01 #6314 @editorialbot
[REVIEW]: Hyperelastics.jl: A Julia package for hyperelastic1 material modelling with a large collection of models
**Submitting author:** @cfarm6 (<a href="http://orcid.org/0000-0002-0456-8301">Carson Farmer</a>)
**Repository:** https://github.com/TRACER-LULab/Hyperelastics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
