### 2024-02-02 TRACER-LULab/Hyperelastics.jl#109 @fb456
(JOSS) Documentation of software installation
I noticed that there is no explicit set of instructions in the docs to install the software. The required libraries can be inferred from the Example section (I managed to install it easily in the REPL pkg mode). A simple section in the docs that outlines this, with a clear list of other dependencies, and maybe a bit of code that automates the installation, would help.
