### 2024-02-02 TRACER-LULab/Hyperelastics.jl#110 @fb456
(JOSS) Statement of Need and Community Guidelines in the Documentation
At the moment, the documentation is missing the Statement of Need and Community Guidelines sections.
