### 2024-02-02 TRACER-LULab/Hyperelastics.jl#111 @fb456
(JOSS) Bug when running Example in the documentation
When trying to run this part of the Example in the docs:
```
models = Dict(
    Gent => ComponentVector(μ=240e-3, Jₘ=80.0),
    EdwardVilgis => ComponentVector(Ns=0.10, Nc=0.20, α=0.001, η=0.001),
    NeoHookean => ComponentVector(μ=200e-3),
    NonaffineMicroSphere => ComponentVector(μ=0.292, N=22.5, p=1.471, U=0.744, q=0.1086),
    Beda => ComponentVector(C1=0.1237, C2=0.0424, C3=7.84e-5, K1=0.0168, α=0.9, β=0.68, ζ=3.015)
)

sol = Dict{Any, SciMLSolution}()
for (ψ, p₀) in models
    HEProblem = HyperelasticProblem(ψ(), treloar_data, p₀,  ad_type = AutoForwardDiff())
    sol[ψ] = solve(HEProblem, NelderMead())
end
```
I get the following error that arises from the Gent model:
```
ERROR: type NamedTuple has no field J_m
Stacktrace:
  [1] getproperty
    @ .\Base.jl:37 [inlined]
  [2] getindex(#unused#::ComponentArrays.Axis{(μ = 1, Jₘ = 2)}, s::Symbol)
    @ ComponentArrays C:\Users\...\.julia\packages\ComponentArrays\NoToB\src\axis.jl:158
  [3] _broadcast_getindex_evalf
    @ .\broadcast.jl:683 [inlined]
  [4] _broadcast_getindex
    @ .\broadcast.jl:656 [inlined]
  [5] (::Base.Broadcast.var"#31#32"{Base.Broadcast.Broadcasted{Base.Broadcast.Style{Tuple}, Nothing, typeof(getindex), Tuple{Tuple{ComponentArrays.Axis{(μ = 1, Jₘ = 2)}}, Base.Broadcast.Broadcasted{Base.Broadcast.Style{Tuple}, Nothing, typeof(ComponentArrays.getval), Tuple{Tuple{DataType}}}}}})(k::Int64)
    @ Base.Broadcast .\broadcast.jl:1088
  [6] ntuple
    @ .\ntuple.jl:48 [inlined]
  [7] copy
    @ .\broadcast.jl:1088 [inlined]
  [8] materialize(bc::Base.Broadcast.Broadcasted{Base.Broadcast.Style{Tuple}, Nothing, typeof(getindex), Tuple{Tuple{ComponentArrays.Axis{(μ = 1, Jₘ = 2)}}, Base.Broadcast.Broadcasted{Base.Broadcast.Style{Tuple}, Nothing, typeof(ComponentArrays.getval), Tuple{Tuple{DataType}}}}})
    @ Base.Broadcast .\broadcast.jl:873
  [9] #s39#55
    @ C:\Users\...\.julia\packages\ComponentArrays\NoToB\src\array_interface.jl:120 [inlined]
 [10] var"#s39#55"(::Any, index_fun::Any, x::Any, idx::Any)
    @ ComponentArrays .\none:0
 [11] (::Core.GeneratedFunctionStub)(::Any, ::Vararg{Any})
    @ Core .\boot.jl:602
 [12] getindex
    @ C:\Users\...\.julia\packages\ComponentArrays\NoToB\src\array_interface.jl:103 [inlined]
 [13] getindex(x::ComponentVector{Float64, Vector{Float64}, Tuple{ComponentArrays.Axis{(μ = 1, Jₘ = 2)}}}, idx::Symbol)
    @ ComponentArrays C:\Users\...\.julia\packages\ComponentArrays\NoToB\src\array_interface.jl:102
 [14] HyperelasticProblem(ψ::Gent{PrincipalValueForm}, test::HyperelasticUniaxialTest{Float64, Float64}, u0::ComponentVector{Float64, Vector{Float64}, Tuple{ComponentArrays.Axis{(μ = 1, J_m = 2)}}}; ad_type::AutoForwardDiff{nothing, Nothing}, loss::LossFunctions.L2DistLoss, lb::NamedTuple{(:μ, :Jₘ), Tuple{Float64,
Float64}}, ub::Nothing, int::Nothing, lcons::Nothing, ucons::Nothing, sense::Nothing, kwargs::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
    @ HyperelasticsOptimizationExt C:\Users\...\.julia\packages\Hyperelastics\9siTz\ext\HyperelasticsOptimizationExt.jl:52
 [15] top-level scope
    @ c:\Users\...\he_review\src\he_review.jl:42
```

Changing "Jₘ" to "J_m" above hasn't helped for this. In either case, I think the issue arises from trying to access a named tuple like this in the HyperelasticsOptimizationExt.jl file:

```
(μ = 0.0, Jₘ = 55.46379855669157)[J_m]
```

The rest of the example runs fine when I remove the Gent model from the code.
