### 2024-02-02 TRACER-LULab/Hyperelastics.jl#112 @fb456
(JOSS) New references in the JOSS Paper Draft
A few additions to the citations would be appropriate, including:
 - on the use of hyperelastic models to model biological materials or polymeric materials (first paragraph of the statement of need)
 - Optimization.jl as it directly used in the fitting
