### 2024-02-02 #6314 @fb456
Hi @cfarm6!

I've also started the review. I have submitted a few issues already that should be addressed, specifically about:
[https://github.com/TRACER-LULab/Hyperelastics.jl/issues/109](url):  software installation
[https://github.com/TRACER-LULab/Hyperelastics.jl/issues/110](url):  Statement of Need and Community Guidelines in the docs
[https://github.com/TRACER-LULab/Hyperelastics.jl/issues/111](url):  A bug I encountered when running the Example
[https://github.com/TRACER-LULab/Hyperelastics.jl/issues/112](url):  A few references that could be added in the paper

I'll continue my review next week and let you know if any more issues arise.
