### 2024-02-02 00dc5e4
Added installation instructions
