### 2024-02-02 TRACER-LULab/Hyperelastics.jl#109 @cfarm6
@fb456  Thank you. Installation instructions have been added.
