### 2024-02-02 TRACER-LULab/Hyperelastics.jl#110 @cfarm6
@fb456 Thank you. This has been addressed.
