### 2024-02-02 TRACER-LULab/Hyperelastics.jl#110 @fb456
Hi. I don't think you've added a Community Guidelines section, something like the link that says:

Community guidelines (https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)
There should be clear guidelines for third-parties wishing to:

Contribute to the software
Report issues or problems with the software
Seek support
