### 2024-02-02 d80f3f0
Added References to the Paper and corrections to contributions
