+++ joss/paper.bib
@@ -5,6 +5,39 @@
   year   = {2023}
 }
 
+@article{wex2015isotropic,
+  title     = {Isotropic incompressible hyperelastic models for modelling the mechanical behaviour of biological tissues: a review},
+  author    = {Wex, Cora and Arndt, Susann and Stoll, Anke and Bruns, Christiane and Kupriyanova, Yuliya},
+  journal   = {Biomedical Engineering/Biomedizinische Technik},
+  volume    = {60},
+  number    = {6},
+  pages     = {577--592},
+  year      = {2015},
+  publisher = {De Gruyter}
+}
+
+@article{beda2014approach,
+  title     = {An approach for hyperelastic model-building and parameters estimation a review of constitutive models},
+  author    = {Beda, T},
+  journal   = {European Polymer Journal},
+  volume    = {50},
+  pages     = {97--108},
+  year      = {2014},
+  publisher = {Elsevier}
+}
+
+@software{vaibhav_kumar_dixit_2023_7738525,
+  author    = {Vaibhav Kumar Dixit and
+               Christopher Rackauckas},
+  title     = {Optimization.jl: A Unified Optimization Package},
+  month     = mar,
+  year      = 2023,
+  publisher = {Zenodo},
+  version   = {v3.12.1},
+  doi       = {10.5281/zenodo.7738525},
+  url       = {https://doi.org/10.5281/zenodo.7738525}
+}
+
 @article{sussman2009model,
   title     = {A model of incompressible isotropic hyperelastic material behavior using spline interpolations of tension--compression test data},
   author    = {Sussman, Theodore and Bathe, Klaus-J{\"u}rgen},
