### 2024-02-02 TRACER-LULab/Hyperelastics.jl#111 @cfarm6
Thank you. I believe this has been fixed now. Please feel free to test any of the code in the documentation. There was an issue with the documentation not updating that has now been resolved.
