### 2024-02-02 TRACER-LULab/Hyperelastics.jl#110 @cfarm6
@fb456 Thank you for the link. I think that the Community Guidelines have now been addressed. Please let me know if there are any changes that you would like to see. It may take a couple minutes for the updated documentation to deploy.
