### 2024-02-02 TRACER-LULab/Hyperelastics.jl#112 @cfarm6
@fb456 Thank you. The appropriate references have been added to the paper.
