+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"example/#Package-Imports","page":"Example","title":"Package Imports","text":"","category":"section"},{"location":"example/","page":"Example","title":"Example","text":"using Hyperelastics\nusing Optimization, OptimizationOptimJL\nusing ComponentArrays: ComponentVector\nusing ForwardDiff\nusing CairoMakie, MakiePublication\nset_theme!(theme_latexfonts())","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"For commonly used datasets in hyperelastic modelling, such as the Treloar1944Uniaxial data, functions are available for getting the datasets:","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"f = Figure(size = (800,800))\nax = Makie.Axis(f[1,1], xlabel = \"Stretch [-]\", ylabel = \"Stress [kg/cm²]\")\ntreloar_data = Treloar1944Uniaxial()\nscatter!(ax, \n    getindex.(treloar_data.data.λ, 1), \n    getindex.(treloar_data.data.s, 1), \n    label = \"Treloar 1944 Experimental\",\n    color = :black\n)\naxislegend(position = :lt)\nsave(\"treloar_data.png\", f)","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"(Image: )","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"Multiple dispatch is used on the corresponding function to calculate the values. Based on the model passed to the function, the correct method will be used in the calculation. StrainEnergyDensity, SecondPiolaKirchoffStressTensor, and CauchyStressTensor accept the deformation state as either the principal components in a vector, [λ₁, λ₂, λ₃] or as the deformation gradient matrix, Fᵢⱼ. The returned value matches the type of the input. Parameters are accessed by field allowing for structs, NamedTuples, or other field-based data-types such as those in ComponentArrays.jl and LabelledArrays.jl. For example, the NeoHookean model is accessed with:","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"ψ = NeoHookean()\nλ⃗ = [2.0, sqrt(1/2), sqrt(1/2)]\np = (μ = 10.0, )\nW = StrainEnergyDensity(ψ, λ⃗, p)\nreturn W # hide","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"or","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"F = rand(3,3)\np = (μ = 20.0, )\nW = StrainEnergyDensity(ψ, F, p)\nreturn W # hide","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"A method for creating an OptimizationProblem compatible with Optimization.jl is provided. To fit the NeoHookean model to the Treloar data previously loaded, an additional field-indexed array is used as the initial guess to HyperelasticProblem. It is recommendedto use ComponentArrays.jl for optimization of model parameters.","category":"page"},{"location":"example/","page":"Example","title":"Example","text":"prob = HyperelasticProblem(ψ, treloar_data, ComponentVector(μ = 0.2), ad_type = AutoForwardDiff())\nsol = solve(prob, LBFGS())\nreturn sol # hide","category":"page"},{"location":"example/","page":"Example","tit
[…]
