### 2024-02-06 TRACER-LULab/Hyperelastics.jl#110 @SotaYoshida
Let me jump in to give a minor comment.
For the first item in the [Community Guidelines](https://tracer-lulab.github.io/Hyperelastics.jl/dev/#Community-Guidelines), the brackets `[]`,`()` used to specify links are reversed.
