### 2024-02-06 TRACER-LULab/Hyperelastics.jl#111 @SotaYoshida
I encountered the same error regarding `J_m`. Could you explain a bit more on how to fix this?
(I assume you are trying to update and deploy the documentation now though...)
