+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Hyperelastics"
 uuid = "c20c605d-adbe-46e4-9db6-e2b04c3ba023"
 authors = ["Carson Farmer"]
-version = "0.1.1"
+version = "0.1.2"
 
 [deps]
 ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
