+++ v0.1.1/paper.bib
@@ -1,0 +1,827 @@
+@article{Bezanson2017,
+  author   = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  title    = {{J}ulia: A Fresh Approach to Numerical Computing},
+  journal  = {SIAM Review},
+  volume   = {59},
+  number   = {1},
+  pages    = {65-98},
+  year     = {2017},
+  doi      = {10.1137/141000671},
+  url      = {
+              https://doi.org/10.1137/141000671
+              },
+  eprint   = {
+              https://doi.org/10.1137/141000671
+              },
+}
+
+@article{mooney1940theory,
+  title     = {A theory of large elastic deformation},
+  author    = {Mooney, Melvin},
+  journal   = {Journal of applied physics},
+  volume    = {11},
+  number    = {9},
+  pages     = {582--592},
+  year      = {1940},
+  publisher = {American Institute of Physics}
+}
+
+
+@article{mansouri2014constitutive,
+  title     = {Constitutive modeling of isotropic hyperelastic materials in an exponential framework using a self-contained approach},
+  author    = {Mansouri, MR and Darijani, H},
+  journal   = {International Journal of Solids and Structures},
+  volume    = {51},
+  number    = {25-26},
+  pages     = {4316--4326},
+  year      = {2014},
+  publisher = {Elsevier}
+}
+
+@article{bahreman2015new,
+  title     = {New polynomial strain energy function; application to rubbery circular cylinders under finite extension and torsion},
+  author    = {Bahreman, Marzieh and Darijani, Hossein},
+  journal   = {Journal of Applied Polymer Science},
+  volume    = {132},
+  number    = {13},
+  year      = {2015},
+  publisher = {Wiley Online Library}
+}
+
+@article{beda2005reconciling,
+  title     = {Reconciling the fundamental phenomenological expression of the strain energy of rubber with established experimental facts},
+  author    = {Beda, T},
+  journal   = {Journal of Polymer Science Part B: Polymer Physics},
+  volume    = {43},
+  number    = {2},
+  pages     = {125--134},
+  year      = {2005},
+  publisher = {Wiley Online Library}
+}
+
+@article{mooney1940theory,
+  title     = {A theory of large elastic deformation},
+  author    = {Mooney, Melvin},
+  journal   = {Journal of applied physics},
+  volume    = {11},
+  number    = {9},
+  pages     = {582--592},
+  year      = {1940},
+  publisher = {American Institute of Physics}
+}
+
+@article{treloar1943elasticity,
+  title     = {The elasticity of a network of long-chain molecules—II},
+  author    = {Treloar, LRG},
+  journal   = {Transactions of the Faraday Society},
+  volume    = {39},
+  pages     = {241--246},
+  year      = {1943},
+  publisher = {Royal Society of Chemistry}
+}
+
+@article{isihara1951statistical,
+  title     = {Statistical theory of rubber-like elasticity. IV.(two-dimensional stretching)},
+  author    = {Isihara, Akira and Hashitsume, Natsuki and Tatibana, Masao},
+  journal   = {The Journal of Chemical Physics},
+  volume    = {19},
+  number    = {12},
+  pages     = {1508--1512},
+  year      = {1951},
+  publisher = {American Institute of Physics}
+}
+
+@article{biderman1958calculation,
+  title   = {Calculation of rubber parts},
+  author  = {Biderman, VL},
+  journal = {Rascheti na prochnost},
+  volume  = {40},
+  year    = {1958}
+}
+
+@article{james1975strain,
+  title     = {Strain energy functions of rubber. I. Characterization of gum vulcanizates},
[…]
