### 2024-02-07 TRACER-LULab/Hyperelastics.jl#111 @cfarm6
@SotaYoshida @fb456 It was an issue with code in the optimization wrapper and calculated the upper and lower bounds for some parameters. The package had originally had Jm instead of J_m. I just released a new version which should have fixed the issue. I had been testing on the dev branch instead of the release. If you update the package, you should received the fix.
