### 2024-02-09 TRACER-LULab/Hyperelastics.jl#111 @SotaYoshida
@cfarm6 The problem is still not resolved.
If I understand correctly, NamedTuples in your code use both `Jₘ` and `J_m` and thereby conflict.
Finally, by unifying all the `J_m` in your package to `Jₘ` (and vice versa), I was able to run the sample code.

By the way, I suggest that the sample code be included in the test code and passed to CI.
