### 2024-02-09 TRACER-LULab/Hyperelastics.jl#111 @cfarm6
@SotaYoshida can you provide the output of `import Pkg; Pkg.status()`? The Hyperelastics version with the patch is v0.1.2.

The example code is being moved into the CI shortly.
