### 2024-02-10 TRACER-LULab/Hyperelastics.jl#111 @SotaYoshida
@cfarm6
> can you provide the output of import Pkg; Pkg.status()?
Yes.
```julia
(Hyperelastics) pkg> status
Project Hyperelastics v0.1.2
Status `~/Downloads/Hyperelastics.jl-main/Project.toml`
  [47edcb42] ADTypes v0.2.6
  [b0b7db55] ComponentArrays v0.15.8
  [3a778109] ContinuumMechanicsBase v0.1.1
  [82cc6244] DataInterpolations v4.6.0
  [ffbed154] DocStringExtensions v0.9.3
  [de52edbc] Integrals v4.1.0
  [b82787f2] InverseLangevinApproximations v0.2.0
  [2ee39098] LabelledArrays v1.15.1
  [30fc2ffe] LossFunctions v0.11.1
⌃ [36348300] OptimizationOptimJL v0.1.14
  [65ce6f38] PackageExtensionCompat v1.0.2
  [1fd47b50] QuadGK v2.9.4
⌅ [731186ca] RecursiveArrayTools v3.3.3
  [189a3867] Reexport v1.2.2
⌃ [0bca4576] SciMLBase v2.15.2
  [276daf66] SpecialFunctions v2.3.1
  [09ab397b] StructArrays v0.6.17
  [37e2e46d] LinearAlgebra
  [10745b16] Statistics v1.9.0
```

In `src/isotropic_incompressible_models.jl`, both `Jₘ` and `J_m` is used.
https://github.com/TRACER-LULab/Hyperelastics.jl/blob/225d19321358214fc27b8e7162f3d72eabcf54cc/src/isotropic_incompressible_models.jl
As mentioned above, I believe unifying them into one or the other resolved the error.
