### 2024-02-12 TRACER-LULab/Hyperelastics.jl#111 @cfarm6
@SotaYoshida I have check that they are unified in the latest release. The docs are generating and running the code without error. Can you try creating a clean environment with the latest version to see if the error persists?
