### 2024-02-13 TRACER-LULab/Hyperelastics.jl#121 @fb456
(JOSS) Paper Contents
Hi,

In my opinion, there is no need to include the documentation Example in the paper's main body, despite it being very useful for the documentation.

I think it would be helpful to add a brief, high-level description of the workflow (in words instead of code, maybe even graphically ), possibly mentioning the types of data being dealt with in the input/output. More emphasis on who the software is directed towards, as well as its scope in terms of what it can do would also be useful. In addition, it could be good to further emphasise the software's differences compared to other similar ones.

In the Example, there are some useful paragraphs in between the code parts, which could be compiled and further developed to address some of the issues above.

Let me know if any of the above is unclear.
