### 2024-02-13 TRACER-LULab/Hyperelastics.jl#111 @cfarm6
@fb456 Thank you.

@SotaYoshida Can you try the latest release in a clean environment? I think the issue should be resolved now.
