### 2024-02-14 TRACER-LULab/Hyperelastics.jl#111 @SotaYoshida
@cfarm6 Yes. I succeeded to run the example code in the doc with a clean environment.
Sorry for bothering you. Now you can close the issue.
