### 2024-02-16 #6314 @SotaYoshida
This package is an excellent effort to implement over 70 hyperelastic material models in the Julia language.
This would help students, newcomers and also experts in the field to work on their own materials.
The paper and documentation are both well written.
I recommend publication of this paper and package.

Minor comments from me are as follows:
- It would be better to highlight the features of the package and the intended audience or users in the paper, as raised in the [issue](https://github.com/TRACER-LULab/Hyperelastics.jl/issues/121#issue-2132223511).
- I suggest authors to describe the assumed and recommended environment (OS and version of Julia guaranteed) of the user in the document. Such information can be found at CI settings and elsewhere, but it would help prevent the readers from using this package in a wrong way.
