### 2024-02-23 #6314 @fb456
Hi,
I also recommend publication provided the issues in the link
 https://github.com/TRACER-LULab/Hyperelastics.jl/issues/121#issue-2132223511
is addressed. This is regarding the general structure of the paper.
