### 2024-03-15 4d07606
Updates to the JOSS paper
