+++ joss/paper.md
@@ -34,158 +34,29 @@ The SEDFs included in this package cover most (if not all) of the available anal
 
 Currently, most commercial finite element codes only offer a limited number, often less than 10, of hyperelastic models which limits the extent to which researchers are able to accurately model a given material. The closest project to `Hyperelastics.jl` is the `matADi` project by Andreas Dutzler [@matAdi2023] which has AD support for 18 material models. 
 
-# Short Example with Code
+# Comparison with Similar Packages
 
-For commonly used datasets in hyperelastic modelling, such as the `Treloar1944Uniaxial` data [@treloar1943elasticity]\autoref{fig:fig1}, functions are available for getting the datasets:
+Similar to the `matADi` project by Andreas Dutzler, `Hyperelastics.jl` provides AD compatible implementations of hyperelastic models. However, `Hyperelastics.jl` provides a significantly larger set of material models and offers the potential for including different compressible material model terms. `matADi` does include models for anisotropic hyperelastic materials which is not support in the current version of `Hyperelastics.jl`. `matADi` is focused on python based implementations while `Hyperelastics.jl` is a pure Julia implementation. 
 
-```julia
-using Hyperelastics
-using Optimization, OptimizationOptimJL
-using ComponentArrays: ComponentVector
-using ForwardDiff
-using CairoMakie, MakiePublication
-set_theme!(theme_web(width = 800))
-f = Figure()
-ax = Axis(f[1,1])
-treloar_data = Treloar1944Uniaxial()
-scatter!(ax, 
-    getindex.(treloar_data.data.λ, 1), 
-    getindex.(treloar_data.data.s, 1), 
-    label = "Treloar 1944 Experimental",
-    color = :black
-)
-axislegend(position = :lt)
-```
+For using material models in a simulation, the leading commercial finite element analysis (FEA) programs, such as Abaqus, Ansys, and LS-DYNA, provide a significantly smaller set of hyperelastic models (often less than 10). Implementing a new model in the commercial programs is often a complex and time-consuming process requiring explicit definition of different partial derivative terms. Furthermore, the model implementations are not compatible between commercial programs. To fill the commercial gap, PolymerFEM [link](www.polymerfem.com) provides a slightly larger set of models for implementation with a focus on visco-elastic and visco-plastic models. From the inital work done with `Hyperelastics.jl`, it is expected that the AD compatibility will allow for a significant improvement in terms of model construction for visco-hyperelastic models. More recently, Wan et al. [@wan2024user] implemented over 70 hyperelastic material models as Abaqus subroutines. The implementation is not AD compatible and is limited to the Abaqus FEA software.
 
-Multiple dispatch is used on the corresponding function to calculate the values. Based on the model passed to the function, the correct method will be used in the calculation. StrainEnergyDensity, SecondPiolaKirchoffStressTensor, and CauchyStressTensor accept the deformation state as either the principal components in a vector, $[\lambda_1, \lambda_2, \lambda_3]$ or as the deformation gradient matrix, $F_{ij}$. The returned value matches the type of the input. Parameters are accessed by field allowing for `structs`, `NamedTuples`, or other field-based data-types such as those in ComponentArrays.jl and LabelledArrays.jl. For example, the NeoHookean model is accessed with:
[…]
