+++ joss/paper.bib
@@ -1,3 +1,13 @@
+@article{wan2024user,
+  title     = {User subroutines platform development for rubber hyperelastic constitutive models and its application in finite element analysis},
+  author    = {Wan, Xianqi and Zhang, Yaru and Zhang, Qiang and Zhang, Liqun and Li, Fanzhu},
+  journal   = {Computational Materials Science},
+  volume    = {237},
+  pages     = {112885},
+  year      = {2024},
+  publisher = {Elsevier}
+}
+
 @software{matAdi2023,
   author = {Andreas Dutzler},
   title  = {matADi: Material Definition with Automatic Differentiation},
