### 2024-03-15 ba7e0e1
Updates to the JOSS paper
