+++ joss/paper.bib
@@ -1,3 +1,30 @@
+@article{dalklint2024simultaneous,
+  title     = {Simultaneous shape and topology optimization of inflatable soft robots},
+  author    = {Dalklint, Anna and Wallin, Mathias and Tortorelli, Daniel},
+  journal   = {Computer Methods in Applied Mechanics and Engineering},
+  volume    = {420},
+  pages     = {116751},
+  year      = {2024},
+  publisher = {Elsevier}
+}
+
+@article{zhu2024comparative,
+  title     = {A comparative study of the entanglement models toward simulating hyperelastic behaviors},
+  author    = {Zhu, Lingrui and Zhan, Lin and Xiao, Rui},
+  journal   = {Journal of Applied Mechanics},
+  volume    = {91},
+  number    = {2},
+  year      = {2024},
+  publisher = {American Society of Mechanical Engineers Digital Collection}
+}
+
+@article{agarwal2024ogden,
+  title     = {An Ogden hyperelastic 3D micromechanical model to depict Poynting effect in brain white matter},
+  author    = {Agarwal, Mohit and Pelegri, Assimina A},
+  journal   = {Heliyon},
+  year      = {2024},
+  publisher = {Elsevier}
+}
 @article{wan2024user,
   title     = {User subroutines platform development for rubber hyperelastic constitutive models and its application in finite element analysis},
   author    = {Wan, Xianqi and Zhang, Yaru and Zhang, Qiang and Zhang, Liqun and Li, Fanzhu},
