+++ joss/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Statement of Need
 
-The development of `Hyperelastics.jl` began as a study of the accuracy for a variety of material models for a set of experimental data. Often, researchers rely on custom implementations of material models and the data fitting process to find material parameters that match their experimental data. Hyperelastic models can well represent the nonlinear stress-deformation behavior of many biological tissues [@wex2015isotropic] as well as engineering polymeric materials [@beda2014approach].
+The development of `Hyperelastics.jl` began as a study of the accuracy for a variety of material models for a set of experimental data. Often, researchers rely on custom implementations of material models and the data fitting process to find material parameters that match their experimental data. Hyperelastic models can well represent the nonlinear stress-deformation behavior of many biological tissues [@wex2015isotropic] as well as engineering polymeric materials [@beda2014approach]. Such material models find applications in developing new materials [@zhu2024comparative], understanding biological tissue mechanical response [@agarwal2024ogden], and soft robotics [@dalklint2024simultaneous].
 
 The SEDFs included in this package cover most (if not all) of the available analytical models from the literature to date, from constitutive to phenomelogical models. Furthermore, a selection of data-driven models are incldued as a starting point for the development of new methods.
 
@@ -44,7 +44,7 @@ The `Hyperelastics.jl` package aims to fill the gap of providing an open-source 
 
 ## Use Case: Fitting a Hyperelastic Model to Experimental Data
 
-The common workflow for modelling a new hyperelastic material is shown in figure \autoref{fig:workflow}. The process commonly begins with the user proving a set of stress-stretch data for a set of uniaxial and/or biaxial tension/compression experiments. The data is loaded into `HyperelasticUniaxialTest` or `HyperelasticBiaxialTest` depending on the nature of the experiment. If the material is assumed to be incompressible, the data for the other principal stretches is calculated. The selection of hyperelastic models is based on multiple dispatch. The primary functions for computations are: `StrainEnergyDensity`, `SecondPiolaKirchoffStressTensor`, and `CauchyStressTensor`. The Second Piola-Kirchoff and Cauchy Stress Tensors are computed by using the partial derivatives of the strain energy density function. By using AD, the implementation of partial derivatives for the majority of the models can be skipped. Furthmore, extensions are provided to load compatible AD functions based on the AD-backend selected. Material models are considered as `structs` and multiple dispatch is used to select the correct equation for the given model. The model parameters are similarly stored in either `structs`, `NamedTuples`, or other field-based data-types such as those in `ComponentArrays.jl` and `LabelledArrays.jl`.
[…]
