### 2024-03-15 TRACER-LULab/Hyperelastics.jl#121 @cfarm6
@fb456 and @SotaYoshida Thank you. I have made updates to the paper to include a Use Case section for going over the highlights of the package and a section for comparisons with similar packages and features in other open-source and commercial programs
