### 2024-03-15 #6314 @cfarm6
@AnjaliSandip The issues have been addressed by adding a section for comparisons with other open-source and commercial packages and a section detailing a common use case for the package.
[paper.pdf](https://github.com/openjournals/joss-reviews/files/14610509/paper.pdf)
