### 2024-03-16 TRACER-LULab/Hyperelastics.jl#121 @fb456
Hi @cfarm6.

Thanks. I think the new version of the paper addresses the issues mentioned above.
