### 2024-03-16 #6314 @fb456
Hi,

I think the issues above have been addressed. One very very minor comment from me is that, referring to the comment above, the information about the recommended version of Julia could also be included in the docs.

Otherwise, I recommend the publication of the paper and the package!
