### 2024-03-16 #6314 @cfarm6
@fb456 I have added statements to the installation instructions in the README and the documentation to reference the compatible Julia versions. In addition, the package is unavailable in older versions as I have only listed compatibility for Julia 1.9+ in the Project.toml file. The updates in the documentation should appear in about 20 minutes once they finish rebuilding.
