### 2024-03-20 e37f5ba
updating the .bib file for missing DOIs
