+++ joss/paper.bib
@@ -39,7 +39,8 @@
   author = {Andreas Dutzler},
   title  = {matADi: Material Definition with Automatic Differentiation},
   url    = {https://github.com/adtzlr/matadi},
-  year   = {2023}
+  year   = {2023},
+  doi = {10.5281/zenodo.5519971}
 }
 
 @article{wex2015isotropic,
@@ -109,6 +110,7 @@
   journal   = {Journal of applied physics},
   volume    = {11},
   number    = {9},
+  doi = {10.1063/1.1712836},
   pages     = {582--592},
   year      = {1940},
   publisher = {American Institute of Physics}
@@ -203,6 +205,7 @@
   author    = {Haines, DW and Wilson, WD},
   journal   = {Journal of the Mechanics and Physics of Solids},
   volume    = {27},
+  doi = {10.1016/0022-5096(79)90034-6},
   number    = {4},
   pages     = {345--360},
   year      = {1979},
@@ -335,6 +338,7 @@
   journal   = {Comptes Rendus Mecanique},
   volume    = {338},
   number    = {1},
+  doi = {10.1016/j.crme.2009.12.007},
   pages     = {3--11},
   year      = {2010},
   publisher = {Elsevier}
