### 2024-03-20 1bc3f72
Updated .bib for JOSS paper
