+++ joss/paper.bib
@@ -5,24 +5,27 @@
   volume    = {420},
   pages     = {116751},
   year      = {2024},
-  publisher = {Elsevier}
+  publisher = {Elsevier},
+  doi = {10.1016/j.cma.2024.116751}
 }
 
 @article{zhu2024comparative,
   title     = {A comparative study of the entanglement models toward simulating hyperelastic behaviors},
+  doi = {10.1115/1.4063348},
   author    = {Zhu, Lingrui and Zhan, Lin and Xiao, Rui},
   journal   = {Journal of Applied Mechanics},
   volume    = {91},
   number    = {2},
   year      = {2024},
   publisher = {American Society of Mechanical Engineers Digital Collection}
-}
-
-@article{agarwal2024ogden,
+  }
+  
+  @article{agarwal2024ogden,
   title     = {An Ogden hyperelastic 3D micromechanical model to depict Poynting effect in brain white matter},
   author    = {Agarwal, Mohit and Pelegri, Assimina A},
   journal   = {Heliyon},
   year      = {2024},
+  doi = {10.1016/j.heliyon.2024.e25379},
   publisher = {Elsevier}
 }
 @article{wan2024user,
@@ -32,7 +35,8 @@
   volume    = {237},
   pages     = {112885},
   year      = {2024},
-  publisher = {Elsevier}
+  publisher = {Elsevier},
+  doi = {10.1016/j.commatsci.2024.112885},
 }
 
 @software{matAdi2023,
@@ -49,6 +53,7 @@
   journal   = {Biomedical Engineering/Biomedizinische Technik},
   volume    = {60},
   number    = {6},
+  doi = {10.1515/bmt-2014-0146},
   pages     = {577--592},
   year      = {2015},
   publisher = {De Gruyter}
@@ -57,6 +62,7 @@
 @article{beda2014approach,
   title     = {An approach for hyperelastic model-building and parameters estimation a review of constitutive models},
   author    = {Beda, T},
+  doi = {10.1016/j.eurpolymj.2013.10.006},
   journal   = {European Polymer Journal},
   volume    = {50},
   pages     = {97--108},
@@ -83,6 +89,7 @@
   volume    = {25},
   number    = {1},
   pages     = {53--63},
+  doi = {10.1002/cnm.1105},
   year      = {2009},
   publisher = {Wiley Online Library}
 }
@@ -123,6 +130,7 @@
   journal   = {International Journal of Solids and Structures},
   volume    = {51},
   number    = {25-26},
+  doi = {10.1016/j.ijsolstr.2014.08.018},
   pages     = {4316--4326},
   year      = {2014},
   publisher = {Elsevier}
@@ -135,12 +143,14 @@
   volume    = {132},
   number    = {13},
   year      = {2015},
+  doi = {10.1002/app.41718},
   publisher = {Wiley Online Library}
 }
 
 @article{beda2005reconciling,
   title     = {Reconciling the fundamental phenomenological expression of the strain energy of rubber with established experimental facts},
   author    = {Beda, T},
+  doi = {10.1002/polb.20308},
   journal   = {Journal of Polymer Science Part B: Polymer Physics},
   volume    = {43},
   number    = {2},
@@ -155,6 +165,7 @@
   journal   = {Journal of applied physics},
   volume    = {11},
   number    = {9},
+  doi = {10.1063/1.1712836},
   pages     = {582--592},
   year      = {1940},
   publisher = {American Institute of Physics}
@@ -166,6 +177,7 @@
   journal   = {Transactions of the Faraday Society},
[…]
