+++ joss/paper.bib
@@ -407,7 +407,7 @@
   author    = {Hart-Smith, L John},
   journal   = {Zeitschrift f{\"u}r angewandte Mathematik und Physik ZAMP},
   volume    = {17},
-  doi = {10.1007/bf01597242}
+  doi = {10.1007/bf01597242},
   number    = {5},
   pages     = {608--626},
   year      = {1966},
@@ -476,7 +476,7 @@
   author    = {Kilian, H-G and Enderle, HF and Unseld, K},
   journal   = {Colloid and Polymer Science},
   volume    = {264},
-  doi = {10.1007/bf01410637}
+  doi = {10.1007/bf01410637},
   number    = {10},
   pages     = {866--876},
   year      = {1986},
@@ -680,7 +680,7 @@
   author    = {Flory, Paul J and Erman, Burak},
   journal   = {Macromolecules},
   volume    = {15},
-  doi = {10.1021/ma00231a022}
+  doi = {10.1021/ma00231a022},
   number    = {3},
   pages     = {800--806},
   year      = {1982},
@@ -850,7 +850,7 @@
   number    = {8},
   pages     = {1784--1797},
   year      = {2013},
-  doi = {10.1016/j.jmps.2013.03.009}
+  doi = {10.1016/j.jmps.2013.03.009},
   publisher = {Elsevier}
 }
 
@@ -860,7 +860,7 @@
   journal   = {Journal of the Mechanics and Physics of Solids},
   volume    = {95},
   pages     = {254--269},
-  doi = {10.1016/j.jmps.2016.05.030}
+  doi = {10.1016/j.jmps.2016.05.030},
   year      = {2016},
   publisher = {Elsevier}
 }
