### 2024-03-20 fdd8ad9
Corrected isihara citation
