+++ joss/paper.bib
@@ -188,7 +188,7 @@
   journal   = {The Journal of Chemical Physics},
   volume    = {19},
   number    = {12},
-  doi = {doi.org/10.1063/1.1748111},
+  doi = {10.1063/1.1748111},
   pages     = {1508--1512},
   year      = {1951},
   publisher = {American Institute of Physics}
