### 2024-03-24 #6314 @AnjaliSandip
@arfon Thanks for letting me know.  I previously thought the author completed this.  After re-reading the editorial guide, I see this is not the case, as you state.

After @cfarm6 releases the new software version and deposits a new archive with Zenodo/Figshare, I will update this thread with the archive DOI.
