### 2024-04-02 #6314 @kyleniemeyer
Hi @cfarm6, I'm doing some final checks before publishing.

- In the paper text, can you remove the redundant "figure" word before the Figure 1 reference?
- In the second paragraph of the Use Case section, please correct "The results are can then be plotted..."
- In the Availability section, it should be "GitHub".
